Renamed anything with "TriggerResponse" to "TriggerOrResponse"
