+++ gramhopper/configuration/common_types.py
@@ -3,5 +3,5 @@ from ..triggers.basic_triggers import BaseTrigger
 from ..responses.basic_responses import BaseResponse
 
 
-TriggerResponse = Union[BaseTrigger, BaseResponse]
-GlobalsDict = Dict[str, TriggerResponse]
+TriggerOrResponse = Union[BaseTrigger, BaseResponse]
+GlobalsDict = Dict[str, TriggerOrResponse]
